add useeffect to make animation work
