+++ components/BookList/BookList.js
@@ -7,8 +7,8 @@ export default function BookList({
   handleToggleAlreadyRead,
   handleToggleBookmark,
   handleToggleCurrentlyReading,
-  updateAnimationBookmark,
   setAnimationActiveAlreadyRead,
+  setAnimationActiveBookmark,
 }) {
   return (
     <StyledList>
@@ -20,8 +20,8 @@ export default function BookList({
             handleToggleAlreadyRead={handleToggleAlreadyRead}
             handleToggleBookmark={handleToggleBookmark}
             handleToggleCurrentlyReading={handleToggleCurrentlyReading}
-            updateAnimationBookmark={updateAnimationBookmark}
             setAnimationActiveAlreadyRead={setAnimationActiveAlreadyRead}
+            setAnimationActiveBookmark={setAnimationActiveBookmark}
           />
         </li>
       ))}
