+++ components/Button/AlreadyReadButton.js
@@ -1,7 +1,27 @@
 import styled from "styled-components";
 import { motion } from "framer-motion";
+import { useEffect } from "react";
 
-export default function AlreadyReadButton({ onToggle, id, isAlreadyRead }) {
+export default function AlreadyReadButton({
+  onToggle,
+  id,
+  isAlreadyRead,
+  booksInfo,
+  setAnimationActiveAlreadyRead,
+}) {
+  useEffect(() => {
+    function updateAnimationAlreadyRead(id) {
+      const currentBook = booksInfo.find((book) => book.id === id);
+      if (currentBook.isAlreadyRead) {
+        setAnimationActiveAlreadyRead(true);
+        setTimeout(() => setAnimationActiveAlreadyRead(false), "1500");
+      } else {
+        setAnimationActiveAlreadyRead(false);
+      }
+    }
+    updateAnimationAlreadyRead(id);
+  }, [isAlreadyRead]);
+
   return (
     <StyledButton
       type="button"
